feat: add custom hook for header
