+++ src/imageLink.ts
@@ -1,0 +1,5 @@
+export const TITLE_IMAGE =
+  'https://cdni.iconscout.com/illustration/premium/thumb/man-walking-in-rain-while-using-smartphone-4981190-4150835.png';
+
+export const ABOUT_IMAGE =
+  'https://cdni.iconscout.com/illustration/premium/thumb/happy-schoolgirl-walk-under-umbrella-at-rainy-season-5683488-4736231.png';
