+++ src/teamInfo.ts
@@ -1,0 +1,19 @@
+export const TMATE1 = [
+  'Roman',
+  'https://cdni.iconscout.com/illustration/premium/thumb/farmer-holding-umbrella-5481939-4591517.png',
+  'https://github.com/rdaniliuk',
+  'https://www.linkedin.com/in/roman-daniliuk-6083b612a/',
+];
+export const TMATE2 = [
+  'Oleksandr',
+  'https://cdni.iconscout.com/illustration/premium/thumb/delivery-in-rain-5402140-4518110.png',
+  'https://cdni.iconscout.com/illustration/premium/thumb/farmer-holding-umbrella-5481939-4591517.png',
+  'https://github.com/Alexander-M-rss',
+  '',
+];
+export const TMATE3 = [
+  'Daniil',
+  'https://cdni.iconscout.com/illustration/premium/thumb/male-engineer-holding-umbrella-5459354-4551753.png',
+  'https://github.com/jaysuno0',
+  '',
+];
